+++ BetterZeeRouter/Handlers/DamageTextHandler.cs
@@ -1,7 +1,0 @@
-﻿namespace BetterZeeRouter;
-
-public sealed class DamageTextHandler : RpcMethodHandler {
-  public override bool Process(ZRoutedRpc.RoutedRPCData routedRpcData) {
-    return false;
-  }
-}